# backend-ttgo/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Overvak - Sistema de monitoreo remoto para personas con Alzheimer a través de un wearable.",
    "description": "Documentación oficial de la API RESTful para el ecosistema **Overvak**. \n\nEste backend gestiona la comunicación entre los dispositivos wearables y la aplicación web. \n\n**Módulos del sistema:**\n1. **Autenticación:** Gestión de cuidadores y tokens.\n2. **Pacientes:** Vinculación de dispositivos y perfiles.\n3. **Datos IoT:** Ingesta de sensores, tales como el MAX30102, GPS (NEO6MV2) y acelerómetro.\n4. **Geocercas:** Algoritmos de punto en polígono para zonas seguras.\n5. **Alertas:** Notificaciones de emergencia.",
    "version": "1.2.1"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Introduce tu token así: Bearer <tu_token>"
    }
  },
  "paths": {
...
    "/api/datos/": {
      "post": {
        "tags": [
          "Datos del paciente"
        ],
        "description": "Recibe datos del sensor. Verifica umbrales y genera alertas automáticas de caídas, inactividad o sensor desconectado.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "dispositivo_id": {
                  "type": "string",
                  "example": "a1b2c3d4-1678886400"
                },
                "frecuencia": {
                  "type": "number",
                  "example": 85
                },
                "oxigeno": {
                  "type": "number",
                  "example": 98
                },
                "latitud": {
                  "type": "number",
                  "example": 19.4326
                },
                "longitud": {
                  "type": "number",
                  "example": -99.1322
                },
                "contacto_cardiaco": {
                  "type": "boolean",
                  "example": true
                },
                "caida_detectada": {
                  "type": "boolean",
                  "example": false
                },
                "inactividad_detectada": {
                  "type": "boolean",
                  "example": false
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Dato procesado correctamente",
            "schema": {
              "$ref": "#/definitions/Dato"
            }
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Faltan datos del sensor o ID del dispositivo"
          },
          "404": {
            "description": "Dispositivo no vinculado a un paciente"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/datos/paciente/{pacienteId}": {
      "get": {
        "tags": [
          "Datos del paciente"
        ],
        "description": "Obtiene el historial completo de datos de un paciente específico",
        "parameters": [
          {
            "name": "pacienteId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del paciente (MongoDB ID)"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de datos obtenida"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Paciente no encontrado o no autorizado"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/datos/reciente/{pacienteId}": {
      "get": {
        "tags": [
          "Datos del paciente"
        ],
        "description": "Obtiene únicamente la última lectura registrada del paciente",
        "parameters": [
          {
            "name": "pacienteId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del paciente"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Último dato obtenido"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Sin datos o paciente no encontrado"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "Dato": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "645d9d..."
        },
        "paciente": {
          "type": "string",
          "example": "645d9c..."
        },
        "frecuencia": {
          "type": "number",
          "example": 85
        },
        "oxigeno": {
          "type": "number",
          "example": 98
        },
        "latitud": {
          "type": "number",
          "example": 19.4326
        },
        "longitud": {
          "type": "number",
          "example": -99.1332
        },
        "contacto_cardiaco": {
          "type": "boolean",
          "example": true
        },
        "caida_detectada": {
          "type": "boolean",
          "example": false
        },
        "inactividad_detectada": {
          "type": "boolean",
          "example": false
        },
        "fecha": {
          "type": "string",
          "example": "2025-05-26T10:00:00.000Z"
        }
      }
    },
...
  }
}